Updated to build and run with XCode 7
While the app will now build and run in XCode 7, however, reverse navigation is broken. So Apple has changed something in iOS 9 that I need to address in the near future.

The new compiler caught an issue with using the numberStyle property, the code was using NumberStyle.

The application available on the App Store runs correctly on iOS 9. Its just the updated libraries that 'break' the application.